+++ lib/pow_assent/phoenix/views/view_helpers.ex
@@ -39,7 +39,7 @@ defmodule PowAssent.Phoenix.ViewHelpers do
   """
   @spec authorization_link(Conn.t(), atom()) :: HTML.safe()
   def authorization_link(conn, provider) do
-    query_params = authorization_link_query_params(conn)
+    query_params = invitation_token_query_params(conn) ++ request_path_query_params(conn)
 
     msg  = AuthorizationController.extension_messages(conn).login_with_provider(%{conn | params: %{"provider" => provider}})
     path = AuthorizationController.routes(conn).path_for(conn, AuthorizationController, :new, [provider], query_params)
@@ -47,8 +47,11 @@ defmodule PowAssent.Phoenix.ViewHelpers do
     Link.link(msg, to: path)
   end
 
-  defp authorization_link_query_params(%{assigns: %{invited_user: %{invitation_token: token}}}), do: [invitation_token: token]
-  defp authorization_link_query_params(_conn), do: []
+  defp invitation_token_query_params(%{assigns: %{invited_user: %{invitation_token: token}}}), do: [invitation_token: token]
+  defp invitation_token_query_params(_conn), do: []
+
+  defp request_path_query_params(%{assigns: %{request_path: request_path}}), do: [request_path: request_path]
+  defp request_path_query_params(_conn), do: []
 
   @doc """
   Generates a provider deauthorization link.